reorganize the metrics folder
