--- checkpoints
@@ -1,1 +1,0 @@
-/home/PJLAB/liqikai/checkpoints/
