+++ mmpose/evaluation/functional/__init__.py
@@ -1,0 +1,10 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+from .keypoint_eval import (keypoint_auc, keypoint_epe, keypoint_nme,
+                            keypoint_pck_accuracy,
+                            multilabel_classification_accuracy,
+                            pose_pck_accuracy)
+
+__all__ = [
+    'keypoint_pck_accuracy', 'keypoint_auc', 'keypoint_nme', 'keypoint_epe',
+    'pose_pck_accuracy', 'multilabel_classification_accuracy'
+]
