+++ mmpose/models/heads/regression_heads/regression_head.py
@@ -8,7 +8,7 @@ from torch import Tensor, nn
 from mmpose.core.utils.tensor_utils import to_numpy
 from mmpose.core.utils.typing import (ConfigType, OptConfigType, OptSampleList,
                                       SampleList)
-from mmpose.metrics.utils import keypoint_pck_accuracy
+from mmpose.evaluation.functional import keypoint_pck_accuracy
 from mmpose.registry import KEYPOINT_CODECS, MODELS
 from ..base_head import BaseHead
